Sentiment program with database
This program determines whether an entered sentence is positive or negative. The entered sentence and its result are saved in the database.
Automation consists of sections called "Creating a Table, Adding Information to the Database, Deleting Data from the Database, Evaluating the Sentence and Ending the Program".This program determines whether an entered sentence is positive or negative. The entered sentence and its result are saved in the database.
Automation consists of sections called "Creating a Table, Adding Information to the Database, Deleting Data from the Database, Evaluating the Sentence and Ending the Program".